+++ docs/CHANGE_LOG.md
@@ -15,6 +15,8 @@
 
 #### **Fixes:**
 - Fixed the binding option "monthsToShow" not being read when set.
+- Fixed the CSS classes "div.day-color-*" being in the opacity order.
+- Fixed a major fault that prevented the heat map showing the other CSS classes for days correctly.
 
 <br>