+++ docs/CHANGE_LOG.md
@@ -1,5 +1,11 @@
 # Heat.js - Change Log:
 
+## Version 0.4.0:
+-
+
+<br>
+
+
 ## Version 0.3.0:
 
 #### **Binding Options:**